moving around lots of files and updating test script
