+++ src/main/java/com/revature/exception/BadCaseException.java
@@ -1,4 +1,4 @@
-package com.revature;
+package com.revature.exception;
 
 public class BadCaseException extends RuntimeException {
     public BadCaseException (String message){
