+++ src/test/java/com/revature/converter/NegativeConverterTest.java
@@ -1,9 +1,11 @@
-package com.revature;
+package com.revature.converter;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import com.revature.exception.BadCaseException;
 
 public class NegativeConverterTest extends ConverterParent{
