+++ src/test/java/com/revature/converter/PositiveConverterTest.java
@@ -1,4 +1,4 @@
-package com.revature;
+package com.revature.converter;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
